Migrated crown profile models to their own module. Resolved any dependencies. Further testing will be implemented.
