+++ fastfuels_core/trees.py
@@ -9,11 +9,14 @@ from fastfuels_core.base import ObjectIterableDataFrame
 from fastfuels_core.point_process import run_point_process
 from fastfuels_core.voxelization import VoxelizedTree, voxelize_tree
 from fastfuels_core.treatments import TreatmentProtocol
+from fastfuels_core.crown_profile_models.abc import CrownProfileModel
+from fastfuels_core.crown_profile_models.purves import PurvesCrownProfile
+from fastfuels_core.crown_profile_models.beta import BetaCrownProfile
 
 # External Imports
 import numpy as np
 from numpy import ndarray
-from scipy.special import beta
+
 from nsvb.estimators import total_foliage_dry_weight
 from pandera import DataFrameSchema, Column, Check, Index
 
@@ -354,242 +357,6 @@ class Tree:
             x=row.X,
             y=row.Y,
         )
-
-
-class CrownProfileModel(ABC):
-    """
-    Abstract base class representing a tree crown profile model.
-    The crown profile model is a rotational solid that can be queried at any
-    height to get a radius, or crown width, at that height.
-
-    This abstract class provides methods to get the radius of the crown at a
-    given height and to get the maximum radius of the crown.
-    """
-
-    @abstractmethod
-    def get_radius_at_height(self, height):
-        """
-        Abstract method to get the radius of the crown at a given height.
-        """
-        raise NotImplementedError(
-            "The get_radius method must be implemented in a subclass."
-        )
-
-    @abstractmethod
-    def get_max_radius(self) -> float:
-        """
-        Abstract method to get the maximum radius of the crown.
-        """
-        raise NotImplementedError(
-            "The get_max_radius method must be implemented in a subclass."
-        )
-
-
-class BetaCrownProfile(CrownProfileModel):
-    """
-    A crown profile model based on a beta distribution.
-    """
-
-    a: float
-    b: float
-    c: float
-    crown_length: float
-    crown_base_height: float
-
-    def __init__(
-        self, species_code: int, crown_base_height: float, crown_length: float
-    ):
-        """
-        Initializes a BetaCrownProfile instance.
-        """
-        self.species_code = int(species_code)
-        self.crown_base_height = crown_base_height
-        self.crown_length = crown_length
-
-        species_group = SPCD_PARAMS[str(self.species_code)]["SPGRP"]
-        self.a = SPGRP_PARAMS[str(species_group)]["BETA_CANOPY_a"]
-        self.b = SPGRP_PARAMS[str(species_group)]["BETA_CANOPY_b"]
-        self.c = SPGRP_PARAMS[str(species_group)]["BETA_CANOPY_c"]
-        self.beta = beta(self.a, self.b)
-
-    def get_max_radius(self) -> float:
-        """
-        Returns the maximum radius of the crown. This function finds the mode
-        of the beta distribution, which is the value of z at which the beta
-        distribution is at its max, and then calculates the radius at that
-        height.
-        """
-        # Find the mode of the beta distribution. This is the value of z at
-        # which the beta distribution is at its max.
-        z_max = (self.a - 1) / (self.a + self.b - 2)
-        normalized_max_radius = self._get_radius_at_normalized_height(z_max)
-        return normalized_max_radius * self.crown_length
-
-    def get_radius_at_height(self, height):
-        """
-        Returns the radius of the crown at a given height using the beta
-        distribution crown model described in equation (3) in Ferrarese et
-        al. (2015). Equation (3) gives the radius of the crown at a given
-        height as a proportion of the crown length scaled between 0 and 1. To
-        get a crown radius in meters, the result is multiplied by the crown
-        length.
-        """
-        normalized_height = self._get_normalized_height(height)
-        radius_at_normalized_height = self._get_radius_at_normalized_height(
-            normalized_height
-        )
-        return radius_at_normalized_height * self.crown_length
-
-    def _get_normalized_height(self, height):
-        """
-        Converts a height (in meters) of the tree crown to a unitless height
-        between 0 and 1 representing the proportion of the crown length.
-        """
-        return (height - self.crown_base_height) / self.crown_length
-
-    def _get_radius_at_normalized_height(self, z):
-        """
-        Returns the unitless radius of the crown at a given normalized height.
-        The radius is scaled between 0 and 1 and represents a proportion of
-        the crown length of the tree.
-
-        The radius is calculated using the beta distribution probability
-        density function (PDF) at the given normalized height. The PDF of the
-        beta distribution uses an additional scaling factor, 'c', for the
-        application to crown profiles and is described by equation (3) in
-        Ferrarese et al. (2015).
-
-        """
-        z = np.asarray(z)
-        mask = (z >= 0) & (z <= 1)
-
-        result = np.zeros_like(z)
-        result[mask] = (
-            self.c * z[mask] ** (self.a - 1) * (1 - z[mask]) ** (self.b - 1) / self.beta
-        )
-
-        if result.size == 1:
-            return result.item()  # Return as a scalar
-        else:
-            return result  # Return as an array
-
-
-class PurvesCrownProfile(CrownProfileModel):
-    """
-    Purves Crown Profile Model.
-
-    This class computes the crown profile for a given tree based on the Purves et al. (2007) model.
-    It uses species-specific parameters to calculate the crown radius at different heights; parameters
-    can be found in fastfuels_core/data/spcd_parameters.json.
-
-    Parameters
-    ----------
-    species_code : int
-        The species code of the tree.
-    dbh : float
-        Diameter at breast height (cm).
-    height : float
-        Total height of the tree (m).
-    crown_ratio : float
-        Ratio of the crown length to the total height of the tree.
-
-    Attributes
-    ----------
-    height : float
-        Total height of the tree (m).
-    crown_ratio : float
-        Ratio of the crown length to the total height of the tree.
-    crown_base_height : float
-        Height at which the crown starts (m).
-    max_crown_radius : float
-        Maximum radius of the crown (m).
-    shape_parameter : float
-        Shape parameter for the crown profile.
-
-    Methods
-    -------
-    get_radius_at_height(height)
-        Computes the crown radius at a given height.
-    get_max_radius()
-        Returns the maximum crown radius.
-    """
-
-    # See Purves et al. (2007) Table S2 in Supporting Information
-    C0_R0 = 0.503
-    C1_R0 = 3.126
-    C0_R40 = 0.5
-    C1_R40 = 10.0
-    C0_B = 0.196
-    C1_B = 0.511
-
-    def __init__(
-        self, species_code: int, dbh: float, height: float, crown_ratio: float
-    ):
-        """
-        Initializes the PurvesCrownProfile model with the given parameters.
-
-        Parameters
-        ----------
-        species_code : int
-            The species code of the tree.
-        dbh : float
-            Diameter at breast height (cm).
-        height : float
-            Total height of the tree (m).
-        crown_ratio : float
-            Ratio of the crown length to the total height of the tree.
-        """
-        self.height = height
-        self.crown_ratio = crown_ratio
-        self.crown_base_height = height - height * crown_ratio
-
-        trait_score = SPCD_PARAMS[str(species_code)]["PURVES_TRAIT_SCORE"]
-
-        # Compute maximum crown radius
-        r0j = (1 - trait_score) * self.C0_R0 + trait_score * self.C1_R0
-        r40j = (1 - trait_score) * self.C0_R40 + trait_score * self.C1_R40
-        self.max_crown_radius = r0j + (r40j - r0j) * (dbh / 40.0)
-
-        # Compute crown shape parameter
-        self.shape_parameter = (1 - trait_score) * self.C0_B + trait_score * self.C1_B
-
-    def get_radius_at_height(self, height: float | np.ndarray) -> float | np.ndarray:
-        """
-        Computes the crown radius at a given height.
-
-        Parameters
-        ----------
-        height : float or np.ndarray
-            Height(s) at which to compute the crown radius (m).
-
-        Returns
-        -------
-        float or np.ndarray
-            Crown radius (m) at the given height(s). Returns a float if input height is a scalar,
-            otherwise returns a numpy array.
-        """
-        height = np.asarray(height)
-        radius = np.where(
-            height < self.crown_base_height,
-            0.0,
-            self.max_crown_radius
-            * (np.minimum((self.height - height) / self.height, 0.95) / 0.95)
-            ** self.shape_parameter,
-        )
-        radius = np.nan_to_num(radius, nan=0.0)
-        return radius if radius.size > 1 else radius.item()
-
-    def get_max_radius(self) -> float | np.ndarray:
-        """
-        Returns the maximum crown radius.
-
-        Returns
-        -------
-        float or np.ndarray
-            Maximum crown radius (m).
-        """
-        return self.get_radius_at_height(self.crown_base_height)
-
 
 class BiomassAllometryModel(ABC):
     """
